+++ frontend/src/components/core/Header/Header.tsx
@@ -16,7 +16,7 @@
  */
 
 import React, { ReactElement, ReactNode } from "react"
-import PageLayoutContext from "src/components/core/PageLayoutContext"
+import AppContext from "src/components/core/AppContext"
 import {
   StyledHeader,
   StyledHeaderDecoration,
@@ -29,7 +29,7 @@ export interface HeaderProps {
 }
 
 function Header({ isStale, children }: HeaderProps): ReactElement {
-  const { wideMode, embedded } = React.useContext(PageLayoutContext)
+  const { wideMode, embedded } = React.useContext(AppContext)
 
   return (
     <StyledHeader
